(fix): fixing tests and refactoring code

+++ src/Utils/Validation/ValidateIndexes.cpp
@@ -1,7 +1,7 @@
 #include "Utils/Validation/ValidateIndexes.hpp"
 
 // The function who calls it GUARANTEES that the string only contains numbers separated by whitespaces
-void ValidateIndexes(const string& data, size_t repositorySize) {
+void ValidateIndexes(const string& data, int repositorySize) {
 	istringstream iss(data);
 	int num;
